uws: fold the per-response booleans into the flags word, drop the node-http-compat flag
Review feedback on the node:http server state:

- Delete HttpFlags::usingNodeHttpCompat. The choice of node:http compat is
  already a template instantiation: setNodeHttpCompat() installs the
  IsNodeHttp=true handler set, so the installed vtable is the single source
  of truth. The two paths that are not templated on IsNodeHttp (socketExtSize
  and HttpResponse::upgrade, which is reached through the C API's runtime
  `int ssl`) read it back via HttpContext::isNodeHttp().

- Move the nine per-connection booleans on HttpResponseData into the existing
  `state` word instead of adding more bool members. `state` was a full uint8_t,
  so it widens to uint32_t and gains HTTP_ANCIENT_REQUEST, HTTP_NO_BODY_STATUS,
  HTTP_CLOSE_DELIMITED and the HTTP_NODE_* bits.

  `state` is reset per response, but four of these describe the connection, so
  the reset is now resetResponseState(), which preserves HTTP_CONNECTION_SCOPED.
  That also makes the "framing flags must not leak into the next response on a
  keep-alive socket" rule structural rather than three hand-written clears.

  uws_res_state() consequently returns a word that no longer fits in a byte
  (HTTP_ANCIENT_REQUEST is set for any HTTP/1.0 request, HTTP_NO_BODY_STATUS for
  any 1xx/204, both on plain Bun.serve), so its Rust mirror widens to u32 too.

- Add HttpParser::isNewline() and reuse it; the empty-line skip before a
  request-line is a do/while now, so the first byte is only loaded once.

- Drop two [[maybe_unused]] attributes that were not needed.

Also fix an unrelated framing bug this uncovered: the http2 allowHTTP1 fallback
dropped the Connection: close that renderNativeHeaders asks for via
AUTO_HEADER_CONN_CLOSE, so a close-delimited response ended the connection
without advertising it. Only the keep-alive line is suppressed for a
close-delimited body now; when the user removed the Connection header
renderNativeHeaders sets no bit and none is written.

+++ src/uws_sys/Response.rs
@@ -982,9 +982,12 @@ pub(crate) type H3Response = crate::h3::Response;
 bitflags::bitflags! {
     /// Non-exhaustive bitset — values may carry
     /// unnamed bit combinations.
+    /// Mirrors `uWS::HttpResponseData::state`. That word is wider than a byte —
+    /// it also carries the response-framing and node:http bits above bit 7 — so
+    /// this must stay `u32` even though only the bits below are named here.
     #[repr(transparent)]
     #[derive(Clone, Copy, PartialEq, Eq)]
-    pub struct State: u8 {
+    pub struct State: u32 {
         const HTTP_STATUS_CALLED               = 1;
         const HTTP_WRITE_CALLED                = 2;
         const HTTP_END_CALLED                  = 4;
